Move init_models() to common.ofa.utils.helper
